src/requirements.txt
flask>=2.0.0
pyYAML >= 5.4.1
PyPDF2==1.26.0
beautifulsoup4 >= 4.9.3
requests >= 2.25.1
mysql-connector-python >= 8.0.25
paramiko >= 2.7.2